+++ NEWS.md
@@ -24,6 +24,9 @@ evaluation process.
     models to an ensemble. This reduces the memory footprint at the cost of IO
     overhead as the models are read from the disk or network when required. By
     default, all models are statically attached to an ensemble.
+    * Added `skip_evaluation_elements` parameter that allows skipping one or
+    more steps of the evaluation process. This is useful if some evaluations are
+    not relevant to you.
 
 
 ## Minor changes: